Move password stack elements into their own component
