+++ client/src/PageLayout/components/PasswordWalkthrough/components/PasswordCreation/PasswordCreation.scss
@@ -1,9 +1,4 @@
 @import "../../../../../GlobalStyles/spacing.scss";
-@import "./colors.scss";
-
-@import "./ArrowButton.scss";
-
-$box-size: 40px;
 
 .StackElement {
   margin: spacing(tight) 0;
@@ -16,17 +11,4 @@ $box-size: 40px;
 
 .PracticeButton {
   padding-right: spacing(tight);
-}
-
-.Circle {
-  display: flex;
-  width: $box-size;
-  height: $box-size;
-  border-radius: spacing(extraLoose);
-}
-
-@each $colorName, $color in $colors {
-  .Circle--#{$colorName} {
-    background: map-get($color, $colorName);
-  }
 }
