+++ client/src/PageLayout/components/PasswordWalkthrough/components/PasswordCreation/PasswordCreation.tsx
@@ -1,16 +1,10 @@
 import * as React from "react";
 
 import autobind from "autobind-decorator";
+import PasswordStack from "../components/PasswordStack";
 import XylophoneContainer from "../../../XylophoneContainer";
 
-import {
-  Button,
-  Icon,
-  Modal,
-  Stack,
-  TextStyle,
-  Tooltip
-} from "@shopify/polaris";
+import { Button, Modal, Tooltip } from "@shopify/polaris";
 
 import "./PasswordCreation.scss";
 
@@ -98,46 +92,13 @@ export default class PasswordCreation extends React.Component<Props, State> {
       </Modal>
     );
 
-    const pageMarkup = [
-      { type: "Shopping", color: "blue", icon: "products" },
-      { type: "Home", color: "teal", icon: "home" },
-      { type: "Phone", color: "red", icon: "notification" }
-    ].map((item, index) => (
-      <div className="StackElement">
-        <Stack alignment="center">
-          <Stack.Item>
-            <div className={`Circle Circle--${item.color}`}>
-              <Icon
-                source={
-                  item.icon === "products"
-                    ? "products"
-                    : item.icon === "home"
-                    ? "home"
-                    : "notification"
-                }
-                color="white"
-              />
-            </div>
-          </Stack.Item>
-          <Stack.Item fill>
-            <TextStyle variation="subdued">{item.type}</TextStyle>
-          </Stack.Item>
-          <Stack.Item>
-            {index === step && (
-              <ul className={`ArrowButton ArrowButton--${item.color}`}>
-                <li>
-                  <a onClick={handleModal}>Create password</a>
-                </li>
-              </ul>
-            )}
-          </Stack.Item>
-        </Stack>
-      </div>
-    ));
-
     return (
       <>
-        {pageMarkup}
+        <PasswordStack
+          step={step}
+          buttonText="Create password"
+          onClick={handleModal}
+        />
         {modalMarkup}
       </>
     );
