+++ client/src/PageLayout/components/PasswordWalkthrough/components/components/PasswordStack/ArrowButton.scss
@@ -1,4 +1,4 @@
-@import "../../../../../GlobalStyles/spacing.scss";
+@import "../../../../../../GlobalStyles/spacing.scss";
 @import "./colors.scss";
 
 $arrow-button-height: 30px;
