+++ client/src/PageLayout/components/PasswordWalkthrough/components/components/PasswordStack/PasswordStack.scss
@@ -1,0 +1,18 @@
+@import "../../../../../../GlobalStyles/spacing.scss";
+@import "./colors.scss";
+@import "./ArrowButton.scss";
+
+$box-size: 40px;
+
+.Circle {
+  display: flex;
+  width: $box-size;
+  height: $box-size;
+  border-radius: spacing(extraLoose);
+}
+
+@each $colorName, $color in $colors {
+  .Circle--#{$colorName} {
+    background: map-get($color, $colorName);
+  }
+}
