+++ client/src/PageLayout/components/PasswordWalkthrough/components/components/PasswordStack/colors.scss
