+++ client/src/PageLayout/components/PasswordWalkthrough/components/components/PasswordStack/index.ts
@@ -1,0 +1,3 @@
+import PasswordStack from "./PasswordStack";
+
+export default PasswordStack;
